requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
google-cloud-bigquery==3.25.0
google-auth==2.35.0
